# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TogetherPlan API",
        "description": "API documentation for the TogetherPlan project. This API allows users to manage events, participants, voting on dates, groups, and notifications.",
        "contact": {
            "name": "TogetherPlan Support",
            "email": "support@togetherplan.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.togetherplan.com",
            "description": "Production server"
        },
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/notifications": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Get unread notifications for the authenticated user",
                "description": "List unread notifications",
                "operationId": "911565ad9aff9ba7eded3f189509e578",
                "responses": {
                    "200": {
                        "description": "List of unread notifications",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "string",
                                                "example": "uuid-notification-id"
                                            },
                                            "type": {
                                                "type": "string",
                                                "example": "App\\\\Notifications\\\\NewEventNotification"
                                            },
                                            "data": {
                                                "type": "object",
                                                "example": {
                                                    "message": "You have a new event invitation"
                                                }
                                            },
                                            "read_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": null,
                                                "nullable": true
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-11-10T10:00:00Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/notifications/{id}/read": {
            "patch": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Mark a notification as read",
                "description": "Mark a notification as read",
                "operationId": "65581bb40eba3764fb900f21495814de",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the notification",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "uuid-notification-id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notification marked as read",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Notification marquée comme lue."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Notification not found or already read",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Notification introuvable ou déjà lue."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/events/{eventId}/invite": {
            "post": {
                "tags": [
                    "Participants"
                ],
                "summary": "Inviter des utilisateurs à un événement privé",
                "description": "Inviter un utilisateur à un événement privé via email",
                "operationId": "fb4f7b8a449e9d588effabddc27b9410",
                "parameters": [
                    {
                        "name": "eventId",
                        "in": "path",
                        "description": "ID de l'événement",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "emails": {
                                        "description": "Liste des emails à inviter",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "email"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Invitation envoyée ou erreurs par email",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "user1@example.com": "Invitation envoyée",
                                    "user2@example.com": "Utilisateur introuvable"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Vous ne pouvez pas inviter des participants"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/events/{eventId}/respond": {
            "post": {
                "tags": [
                    "Participants"
                ],
                "summary": "Répondre à une invitation à un événement",
                "description": "Répondre à une invitation (accept/refuse)",
                "operationId": "c357e712cb22566529c8ab01a5e3d3ba",
                "parameters": [
                    {
                        "name": "eventId",
                        "in": "path",
                        "description": "ID de l'événement",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "description": "Statut de la réponse",
                                        "type": "string",
                                        "enum": [
                                            "accepted",
                                            "declined"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Réponse envoyée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invitation accepted avec succès."
                                        },
                                        "participant": {
                                            "$ref": "#/components/schemas/Participant"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Vous n'êtes pas invité à cet événement"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/events/{eventId}/participants": {
            "get": {
                "tags": [
                    "Participants"
                ],
                "summary": "Lister les participants d’un événement",
                "description": "Lister les participants d’un événement",
                "operationId": "d0dafe7f1c94647911f5b26c378b75d3",
                "parameters": [
                    {
                        "name": "eventId",
                        "in": "path",
                        "description": "ID de l'événement",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des participants",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Participant"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès refusé à cet événement privé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "DateOption": {
                "title": "DateOption",
                "description": "Schema representing a proposed date option for an event",
                "required": [
                    "event_id",
                    "proposed_date"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "event_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "proposed_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-11-15"
                    },
                    "proposed_time": {
                        "type": "string",
                        "format": "time",
                        "example": "14:30:00",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-11-03T16:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-11-03T16:05:00Z"
                    },
                    "event": {
                        "$ref": "#/components/schemas/Event"
                    },
                    "votes": {
                        "description": "Votes associated with this date option",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Vote"
                        }
                    }
                },
                "type": "object"
            },
            "Event": {
                "title": "Event",
                "description": "Schema representing an event",
                "required": [
                    "id",
                    "title",
                    "location",
                    "category",
                    "privacy",
                    "created_by"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Team Meeting"
                    },
                    "description": {
                        "type": "string",
                        "example": "Discuss project progress"
                    },
                    "location": {
                        "type": "string",
                        "example": "Casablanca"
                    },
                    "category": {
                        "type": "string",
                        "example": "Meeting"
                    },
                    "privacy": {
                        "type": "string",
                        "enum": [
                            "public",
                            "private"
                        ],
                        "example": "private"
                    },
                    "created_by": {
                        "type": "integer",
                        "example": 3
                    },
                    "group_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "best_date_id": {
                        "type": "integer",
                        "example": 10,
                        "nullable": true
                    },
                    "creator": {
                        "$ref": "#/components/schemas/User"
                    },
                    "group": {
                        "$ref": "#/components/schemas/Group"
                    },
                    "participants": {
                        "description": "List of participants for this event",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Participant"
                        }
                    },
                    "dateOptions": {
                        "description": "List of proposed date options for this event",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/DateOption"
                        }
                    },
                    "bestDate": {
                        "$ref": "#/components/schemas/DateOption"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-11-03T15:50:50Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-11-03T15:50:50Z"
                    }
                },
                "type": "object"
            },
            "Group": {
                "title": "Group",
                "required": [
                    "id",
                    "name",
                    "created_by"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Groupe Dev"
                    },
                    "description": {
                        "type": "string",
                        "example": "Description du groupe"
                    },
                    "created_by": {
                        "type": "integer",
                        "example": 1
                    },
                    "users": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    },
                    "events": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "Participant": {
                "title": "Participant",
                "description": "Schema representing a participant in an event",
                "properties": {
                    "id": {
                        "description": "ID du participant",
                        "type": "integer",
                        "example": 1
                    },
                    "event_id": {
                        "description": "ID de l'événement auquel le participant est associé",
                        "type": "integer",
                        "example": 5
                    },
                    "user_id": {
                        "description": "ID de l'utilisateur participant",
                        "type": "integer",
                        "example": 10
                    },
                    "status": {
                        "description": "Statut de l'invitation",
                        "type": "string",
                        "enum": [
                            "invited",
                            "accepted",
                            "declined"
                        ],
                        "example": "invited"
                    },
                    "created_at": {
                        "description": "Date de création du participant",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-11-03T16:00:00Z"
                    },
                    "updated_at": {
                        "description": "Date de dernière mise à jour du participant",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-11-03T16:05:00Z"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "event": {
                        "$ref": "#/components/schemas/Event"
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User",
                "description": "Schema representing a user",
                "required": [
                    "id",
                    "name",
                    "email"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Zakaria Ghazi"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "zakariaghazi77@gmail.com"
                    },
                    "email_verified_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-11-03T15:50:50Z",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-11-03T15:50:50Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-11-03T15:50:50Z"
                    },
                    "groups": {
                        "description": "Groups the user belongs to",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Group"
                        }
                    },
                    "participants": {
                        "description": "Participant records for this user",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Participant"
                        }
                    },
                    "events_created": {
                        "description": "Events created by this user",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Event"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Use Laravel Sanctum API token for authentication",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}